+++ workers/ohsome_quality_analyst/indicators/ghs_pop_comparison_buildings/indicator.py
@@ -51,8 +51,7 @@ class GhsPopComparisonBuildings(BaseIndicator):
 
     async def preprocess(self) -> None:
         raster = get_raster_dataset("GHS_POP_R2019A")
-        stats = get_zonal_stats(self.feature, raster, stats="sum")
-        pop_count = stats[0]["sum"]
+        pop_count = get_zonal_stats(self.feature, raster, stats="sum")[0]["sum"]
         area = await get_area_of_bpolys(self.feature.geometry)
 
         if pop_count is None: